Refactor: Improve mention input, typography, and component exports
