+++ src/components/ui/typography.tsx
@@ -250,21 +250,3 @@ export const Quote = ({
   </Component>
 );
 
-// Export individual components
-export {
-  H1,
-  H2,
-  H3,
-  H4,
-  H5,
-  H6,
-  BodyLarge,
-  Body,
-  BodySmall,
-  Caption,
-  Label,
-  Display,
-  Code,
-  Quote
-};
-
