+++ src/pages/Notifications.tsx
@@ -9,7 +9,7 @@ import {
   Trash2,
   X,
 } from 'lucide-react';
-import { Typography } from '../components/ui/typography';
+import { Typography } from '../components/ui/typography-utils';
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { LoadingSpinner } from '../components/common/LoadingSpinner';
