+++ src/pages/TodoHome.tsx
@@ -11,7 +11,7 @@ import {
   User as UserIcon,
   Users,
 } from 'lucide-react';
-import { Typography } from '../components/ui/typography';
+import { Typography } from '../components/ui/typography-utils';
 import { useEffect, useMemo, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Calendar as CalendarComponent } from '../components/calendar';
